+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.1+2
 * Check if Widget is mounted before validating
+* Get ErrorColor from ColorScheme
 ## 0.0.1+1
 * Fix Readme