+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.0.1+2
 * Check if Widget is mounted before validating
 * Get ErrorColor from ColorScheme
+* Adjust German Translations
+
 ## 0.0.1+1
 * Fix Readme